A simple folium test script to make the data easy to bind an interactive
The following script uses Folium and Leaflet.js to make interactive
google maps type maps.
I've bound the price data to the map and included some comments so that you
know how this is done.
This might be a good avenue to follow if you are trying to make interactive maps
If you're interested in collaborating let me know, contact info in the script.

Hope this is helpful